Fix structure Factory
